+++ CHANGELOG.rst
@@ -2,9 +2,16 @@
 Changelog
 =========
 
+v1.0.0
+------
+- now has a full test suite! 🎉
+- can now initialize a :code:`NextGenPump` instance using either a pre-existing :code:`Serial` instance (new behavior), OR by passing in a string for the port to open a :code:`Serial` instance at (old behavior)
+- bundled data retrieval methods such as :code:`current_conditions` or :code:`pump_info` now return custom typed :code:`dataclass`es instead of :code:`dict`s (this is a breaking change)
+- fixed an elusive bug that would sometimes cause :code:`NextGenPumpBase`'s :code:`identify` to fail
+
 v0.1.7
 ------
- - minor code cleanup / improvement to packaging
+- minor code cleanup / improvement to packaging
 
 v0.1.6
 ------